Move inputhex component into ui-components
